Renaming __test__ to test. Double underscores is Jest naming convention for the folder containing tests. We decided to put tests aside to source code so this naming doesn't make sense anymore. Use - instead of _ for uniformity.
