+++ package.json
@@ -5,13 +5,13 @@
   "scripts": {
     "build": "npm run clean && npm run copy-assets && run-p build:**",
     "build:html": "parcel build --no-cache public/*.html",
-    "build:scripts": "parcel build --no-cache src/application/background_script.ts",
+    "build:scripts": "parcel build --no-cache src/application/background-script.ts",
     "clean": "rimraf dist",
     "copy-assets": "mkdir -p dist && cp public/manifest.json dist/ && cp -r public/assets dist/",
-    "lint": "eslint --ext .ts,.tsx src __test__",
-    "lint:fix": "eslint --fix --ext .ts,.tsx src __test__",
-    "prettier": "prettier --write src __test__",
-    "prettier:check": "prettier --check src __test__",
+    "lint": "eslint --ext .ts,.tsx src test",
+    "lint:fix": "eslint --fix --ext .ts,.tsx src test",
+    "prettier": "prettier --write src test",
+    "prettier:check": "prettier --check src test",
     "run:firefox": "web-ext run --verbose --browser-console --firefox=firefoxdeveloperedition --firefox-profile=dev --source-dir dist",
     "run:brave-linux": "web-ext run --target chromium --chromium-binary '/usr/bin/brave' --source-dir dist",
     "run:chrome-linux": "web-ext run --target chromium --chromium-binary chromium --source-dir dist",
@@ -34,7 +34,7 @@
     "watchskip": "npm run clean && npm run copy-assets && SKIP_ONBOARDING=true run-p watch:**",
     "watch": "npm run clean && npm run copy-assets && run-p watch:**",
     "watch:html": "parcel watch public/*.html",
-    "watch:scripts": "parcel watch src/application/background_script.ts",
+    "watch:scripts": "parcel watch src/application/background-script.ts",
     "web-ext:build": "web-ext build --source-dir dist",
     "web-ext:lint": "web-ext lint --source-dir dist"
   },
