+++ src/application/background-script.ts
