+++ src/domain/wallet/value-objects/master-blinding-key.test.ts
@@ -1,5 +1,5 @@
 import { MasterBlindingKey } from './master-blinding-key';
-import { masterBlindingKey } from '../../../../__test__/fixtures/wallet.json';
+import { masterBlindingKey } from '../../../../test/fixtures/wallet.json';
 
 test('Should be able to create a master blinding key', () => {
   const res = MasterBlindingKey.create(masterBlindingKey);
