+++ test/fixtures/assets.json
