+++ test/fixtures/tx-interface/receive/receive-lbtc-fee-lbtc.json
