+++ test/fixtures/tx-interface/receive/receive-usdt-fee-lbtc.json
