+++ test/fixtures/tx-interface/send/send-lbtc-fee-lbtc-conf.json
