+++ test/fixtures/tx-interface/send/send-usdt-fee-lbtc-conf.json
